move get_sync_logs_for_user to phone.analytics
